+++ clang/test/CodeGen/functions.c
@@ -16,9 +16,6 @@ void test3(T f) {
   f();
 }
 
-int a(int);
-int a() {return 1;}
-
 void f0(void) {}
 // CHECK-LABEL: define{{.*}} void @f0()